+++ resolwe/flow/executors/docker/constants.py
@@ -4,5 +4,7 @@ DATA_ALL_VOLUME = '/data_all'
 UPLOAD_VOLUME = '/upload'
 SECRETS_VOLUME = '/secrets'
 
+TMPDIR = '.tmp'
+
 # Container image used if no image is specified.
 DEFAULT_CONTAINER_IMAGE = 'resolwe/base:ubuntu-16.04'